[CodeQuality] Rename SimplerWithEqualToRector to CallbackSingleAssertToSimplerRector
